+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to ds-kaltura will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [5.0.0](https://github.com/kb-dk/ds-kaltura/releases/tag/ds-kaltura-5.0.0) - 2026-06-10
+## [Unreleased]
 
 ### Changed
 
 - Removed all scripts, jobs and alot of obsolete methods and tests.
+- Removed all scripts, jobs and a lot of obsolete methods and tests.
+- Use null safe `StringUtils.isBlank` instead of `String.isEmpty`.
+- Start on unittest for `DsKalturaAnalytics`.
+- Refactored creating search query and call to Kaltura from method `listEntryBatch` into own method `getMediaEntries`.
+- Changed method `listEntryBatch` from `private` to `public` to be able to unittest the behavior.
+
+## [5.0.0](https://github.com/kb-dk/ds-kaltura/releases/tag/ds-kaltura-5.0.0) - 2026-06-10
 
 ### Added